+++ packages/react/src/components/index.tsx
@@ -1,4 +1,3 @@
 export * from './modal';
 export * from './tab';
-export * from './carousel';
-export * from './popover';
+export * from './popover1';
