+++ packages/react/src/index.ts
@@ -1,3 +1,4 @@
 export * from './components'
 export * from './elements'
-export * from './layout'
+export * from './layout';
+export  * from './components/carousel'
